Refactor to use fewer unwrap and casts

See merge request surfer-project/surfer!232

+++ src/time.rs
@@ -333,7 +333,7 @@ impl State {
                                 &self.wanted_timeunit,
                                 &self.get_time_format(),
                             ),
-                            viewport.from_time(&tick, frame_width as f64) as f32,
+                            viewport.from_time(&tick, frame_width),
                         )
                     })
                     .collect::<Vec<(String, f32)>>();
